implementations/preprocessment/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pypi_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
ca-certificates=2023.7.22=h56e8100_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
charset-normalizer=3.2.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
cudatoolkit=11.2.2=h933977f_10
cudnn=8.1.0.77=h3e0f4f4_0
cycler=0.11.0=pypi_0
debugpy=1.6.3=py37hf2a7229_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
flatbuffers=23.5.26=pypi_0
fonttools=4.38.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.23.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.58.0=pypi_0
h5py=3.8.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.7.0=pypi_0
ipykernel=6.16.2=pyh025b116_0
ipython=7.33.0=py37h03978a9_0
jedi=0.19.0=pyhd8ed1ab_0
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_core=4.11.1=py37h03978a9_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
libclang=16.0.6=pypi_0
libsodium=1.0.18=h8d14728_1
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numpy=1.21.6=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.76=pypi_0
openssl=1.1.1w=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=22.3.1=py37haa95532_0
prompt-toolkit=3.0.39=pyha770c72_0
protobuf=3.19.6=pypi_0
psutil=5.9.1=py37hcc03f2d_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
python=3.7.16=h6244533_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pywin32=303=py37hcc03f2d_0
pyzmq=23.2.0=py37hcce574b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.7.3=pypi_0
setuptools=65.6.3=py37haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h2bbff1b_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.10.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
tornado=6.2=py37hcc03f2d_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.7.1=pypi_0
urllib3=1.26.16=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37haa95532_0
wincertstore=0.2=py37haa95532_2
wrapt=1.15.0=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.15.0=pypi_0
